+++ tests/conftest.py
@@ -1,11 +1,16 @@
 import os
-import yaml
 
+from qgis.core import QgsSettings
+
+from ORStools.ORStoolsPlugin import ORStools
 from ORStools.utils.configmanager import read_config
+from tests.utils.utilities import get_qgis_app
 
-with open("ORStools/config.yml", "r+") as file:
-    data = yaml.safe_load(file)
+QGISAPP, CANVAS, IFACE, PARENT = get_qgis_app()
 
+ORStools(IFACE).add_default_provider_to_settings()
+s = QgsSettings()
+data = s.value("ORStools/config")
 
 def pytest_sessionstart(session):
     """
@@ -14,8 +19,7 @@ def pytest_sessionstart(session):
     """
     if data["providers"][0]["key"] == "":
         data["providers"][0]["key"] = os.environ.get("ORS_API_KEY")
-        with open("ORStools/config.yml", "w") as file:
-            yaml.dump(data, file)
+        s.setValue("ORStools/config", data)
     else:
         raise ValueError("API key is not empty.")
 
@@ -25,10 +29,8 @@ def pytest_sessionfinish(session, exitstatus):
     Called after whole test run finished, right before
     returning the exit status to the system.
     """
-    with open("ORStools/config.yml", "w") as file:
-        if not data["providers"][0]["key"] == "":
-            data['providers'][0]['key'] = ''  # fmt: skip
-        yaml.dump(data, file)
-
+    if not data["providers"][0]["key"] == "":
+        data['providers'][0]['key'] = ''  # fmt: skip
+    s.setValue("ORStools/config", data)
     config = read_config()
     assert config["providers"][0]["key"] == ''  # fmt: skip